backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.7
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
prometheus-fastapi-instrumentator==7.1.0
llama-cpp-python==0.2.70
httpx==0.25.2

# Redis and caching
redis==5.0.1

# Rate limiting and retries
tenacity==8.2.3
aiobreaker==1.2.0

# Background tasks and queuing
rq==1.15.1
celery==5.3.4

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0
